Settings page cleanup

+++ plugin/admin/options/option-types/class-string-field.php
@@ -5,10 +5,12 @@ require_once __DIR__ . '/class-setting-field.php';
 
 class StringField extends SettingField {
   private $readOnly;
+  private $selectable;
 
   public function __construct( $id, $title, $page, $section, $defaultValue ) {
     parent::__construct( $id, $title, $page, $section, $defaultValue );
     $this->readOnly = false;
+    $this->selectable = false;
   }
 
   public function register() {
@@ -19,13 +21,16 @@ class StringField extends SettingField {
     return esc_html( $value );
   }
 
-  public function addField( $value = false ) {
-    $this->readOnly = $value;
+  public function addField( $read = false, $sel = false ) {
+    $this->readOnly = $read;
+    $this->selectable = $sel;
     parent::addField();
   }
 
   public function display() {
-    if ( $this->readOnly ) {
+    if ( $this->readOnly && $this->selectable ) {
+      echo "<input onClick='this.select();' type='text' name='" . esc_attr( $this->id ) . "' value='" . esc_attr( get_option( $this->id, $this->defaultValue ) ) . "' readonly class='regular-text'>";
+    } else if ( $this->readOnly && !$this->selectable ) {
       echo "<input type='text' name='" . esc_attr( $this->id ) . "' value='" . esc_attr( get_option( $this->id, $this->defaultValue ) ) . "' readonly class='regular-text'>";
     } else {
       echo "<input type='text' name='" . esc_attr( $this->id ) . "' value='" . esc_attr( get_option( $this->id, $this->defaultValue ) ) . "' class='regular-text'>";
